refactor(app/src): Refactor file structure to keep sample files, but rename package to com.android.unio.
