+++ app/src/androidTest/java/com/android/unio/screen/MainScreen.kt
@@ -1,7 +1,7 @@
-package com.android.sample.screen
+package com.android.unio.screen
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
-import com.android.sample.resources.C
+import com.android.unio.resources.C
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
