+++ app/src/main/java/com/android/unio/MainActivity.kt
@@ -7,11 +7,13 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
-import com.android.sample.resources.C
-import com.android.sample.ui.theme.SampleAppTheme
+import androidx.compose.ui.tooling.preview.Preview
+import com.android.unio.resources.C
+import com.android.unio.ui.theme.SampleAppTheme
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
@@ -22,9 +24,20 @@ class MainActivity : ComponentActivity() {
         Surface(
             modifier = Modifier.fillMaxSize().semantics { testTag = C.Tag.main_screen_container },
             color = MaterialTheme.colorScheme.background) {
-              Text("Unio App")
+              Greeting("Android")
             }
       }
     }
   }
+}
+
+@Composable
+fun Greeting(name: String, modifier: Modifier = Modifier) {
+  Text(text = "Hello $name!", modifier = modifier.semantics { testTag = C.Tag.greeting })
+}
+
+@Preview(showBackground = true)
+@Composable
+fun GreetingPreview() {
+  SampleAppTheme { Greeting("Android") }
 }
