+++ app/src/main/java/com/android/unio/SimpleData.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.unio
 
 import kotlin.math.sqrt
