+++ app/src/main/java/com/android/unio/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.android.sample.ui.theme
+package com.android.unio.ui.theme
 
 import android.app.Activity
 import android.os.Build
