+++ app/src/test/java/com/android/unio/PointTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.android.unio
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
